Base refactor
