+++ packages/react-native-reanimated/src/ConfigHelper.native.ts
@@ -5,17 +5,10 @@ import { runOnUISync } from 'react-native-worklets';
 import type { LoggerConfig } from './common';
 import { getLoggerConfig, IS_JEST, updateLoggerConfig } from './common';
 
-/** @deprecated This function is a no-op in Reanimated 4. */
-export function addWhitelistedNativeProps(
-  _props: Record<string, boolean>
-): void {
-  // Do nothing. This is just for backward compatibility.
-}
-
-/** @deprecated This function is a no-op in Reanimated 4. */
-export function addWhitelistedUIProps(_props: Record<string, boolean>): void {
-  // Do nothing. This is just for backward compatibility.
-}
+export {
+  addWhitelistedNativeProps,
+  addWhitelistedUIProps,
+} from './ConfigHelperBase';
 
 /**
  * Updates Reanimated logger config with the user-provided configuration. Will
