+++ packages/react-native-reanimated/src/ConfigHelper.ts
@@ -3,17 +3,10 @@
 import type { LoggerConfig } from './common';
 import { getLoggerConfig, updateLoggerConfig } from './common';
 
-/** @deprecated This function is a no-op in Reanimated 4. */
-export function addWhitelistedNativeProps(
-  _props: Record<string, boolean>
-): void {
-  // Do nothing. This is just for backward compatibility.
-}
-
-/** @deprecated This function is a no-op in Reanimated 4. */
-export function addWhitelistedUIProps(_props: Record<string, boolean>): void {
-  // Do nothing. This is just for backward compatibility.
-}
+export {
+  addWhitelistedNativeProps,
+  addWhitelistedUIProps,
+} from './ConfigHelperBase';
 
 /**
  * Updates Reanimated logger config with the user-provided configuration. Will
