+++ packages/react-native-reanimated/src/ReducedMotion.native.ts
@@ -1,5 +1,5 @@
 'use strict';
-import { makeMutable } from './mutables';
+import { createReducedMotionManager } from './ReducedMotionBase';
 
 type localGlobal = typeof global & Record<string, unknown>;
 
@@ -7,13 +7,6 @@ export function isReducedMotionEnabledInSystem() {
   return !!(global as localGlobal)._REANIMATED_IS_REDUCED_MOTION;
 }
 
-const IS_REDUCED_MOTION_ENABLED_IN_SYSTEM = isReducedMotionEnabledInSystem();
-
-export const ReducedMotionManager = {
-  jsValue: IS_REDUCED_MOTION_ENABLED_IN_SYSTEM,
-  uiValue: makeMutable(IS_REDUCED_MOTION_ENABLED_IN_SYSTEM),
-  setEnabled(value: boolean) {
-    ReducedMotionManager.jsValue = value;
-    ReducedMotionManager.uiValue.value = value;
-  },
-};
+export const ReducedMotionManager = createReducedMotionManager(
+  isReducedMotionEnabledInSystem()
+);
