+++ packages/react-native-reanimated/src/createAnimatedComponent/JSPropsUpdaterBase.ts
@@ -1,0 +1,32 @@
+'use strict';
+import type {
+  AnimatedComponentProps,
+  IAnimatedComponentInternal,
+  IJSPropsUpdater,
+  InitialComponentProps,
+  JSPropsOperation,
+} from './commonTypes';
+
+export class JSPropsUpdaterWeb implements IJSPropsUpdater {
+  public registerComponent(
+    _animatedComponent: React.Component<
+      AnimatedComponentProps<InitialComponentProps>
+    > &
+      IAnimatedComponentInternal
+  ) {
+    // noop
+  }
+
+  public unregisterComponent(
+    _animatedComponent: React.Component<
+      AnimatedComponentProps<InitialComponentProps>
+    > &
+      IAnimatedComponentInternal
+  ) {
+    // noop
+  }
+
+  public updateProps(_operations: JSPropsOperation[]) {
+    // noop
+  }
+}
