+++ packages/react-native-reanimated/src/updateProps/updateProps.native.ts
@@ -1,6 +1,5 @@
 'use strict';
 
-import type { RefObject } from 'react';
 import { scheduleOnRN, scheduleOnUI } from 'react-native-worklets';
 
 import {
@@ -11,20 +10,17 @@ import {
   stylePropsBuilder,
 } from '../common';
 import { processBoxShadowWeb, processFilterWeb } from '../common/web';
-import type {
-  AnimatedStyle,
-  ShadowNodeWrapper,
-  StyleProps,
-} from '../commonTypes';
+import type { ShadowNodeWrapper, StyleProps } from '../commonTypes';
 import type {
   JSPropsOperation,
   PropUpdates,
 } from '../createAnimatedComponent/commonTypes';
 import jsPropsUpdater from '../createAnimatedComponent/JSPropsUpdater';
 import { getStaticFeatureFlag } from '../featureFlags';
-import type { Descriptor } from '../hook/commonTypes';
 import type { ReanimatedHTMLElement } from '../ReanimatedModule/js-reanimated';
 import { _updatePropsJS } from '../ReanimatedModule/js-reanimated';
+import type { ViewDescriptorsWrapper } from './updatePropsBase';
+import { makeUpdatePropsJestWrapper, maybeThrowError } from './updatePropsBase';
 
 const USE_ANIMATION_BACKEND = getStaticFeatureFlag('USE_ANIMATION_BACKEND');
 
@@ -77,25 +73,7 @@ if (IS_JEST) {
   };
 }
 
-export const updatePropsJestWrapper = (
-  viewDescriptors: ViewDescriptorsWrapper,
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  updates: AnimatedStyle<any>,
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  animatedValues: RefObject<AnimatedStyle<any>>,
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  adapters: ((updates: AnimatedStyle<any>) => void)[]
-): void => {
-  adapters.forEach((adapter) => {
-    adapter(updates);
-  });
-  animatedValues.current.value = {
-    ...animatedValues.current.value,
-    ...updates,
-  };
-
-  updateProps(viewDescriptors, updates);
-};
+export const updatePropsJestWrapper = makeUpdatePropsJestWrapper(updateProps);
 
 export default updateProps;
 
@@ -178,15 +156,6 @@ function createUpdatePropsManager() {
 
 // is-tree-shakable-suppress
 if (IS_JEST) {
-  const maybeThrowError = () => {
-    // Jest attempts to access a property of this object to check if it is a Jest mock
-    // so we can't throw an error in the getter.
-    if (!IS_JEST) {
-      throw new Error(
-        '[Reanimated] `UpdatePropsManager` is not available on non-native platform.'
-      );
-    }
-  };
   global.UpdatePropsManager = new Proxy(
     {},
     {
@@ -202,12 +171,4 @@ if (IS_JEST) {
     'worklet';
     global.UpdatePropsManager = createUpdatePropsManager();
   });
-}
-
-/**
- * This used to be `SharedValue<Descriptors[]>` but objects holding just a
- * single `value` prop are fine too.
- */
-interface ViewDescriptorsWrapper {
-  value: Readonly<Descriptor[]>;
 }
