+++ packages/react-native-reanimated/src/updateProps/updateProps.ts
@@ -1,14 +1,11 @@
 'use strict';
 
-import type { RefObject } from 'react';
-
-import { IS_JEST } from '../common';
 import { processBoxShadowWeb, processFilterWeb } from '../common/web';
-import type { AnimatedStyle } from '../commonTypes';
 import type { PropUpdates } from '../createAnimatedComponent/commonTypes';
-import type { Descriptor } from '../hook/commonTypes';
 import type { ReanimatedHTMLElement } from '../ReanimatedModule/js-reanimated';
 import { _updatePropsJS } from '../ReanimatedModule/js-reanimated';
+import type { ViewDescriptorsWrapper } from './updatePropsBase';
+import { makeUpdatePropsJestWrapper, maybeThrowError } from './updatePropsBase';
 
 const updateProps: (
   viewDescriptors: ViewDescriptorsWrapper,
@@ -28,38 +25,10 @@ const updateProps: (
   });
 };
 
-export const updatePropsJestWrapper = (
-  viewDescriptors: ViewDescriptorsWrapper,
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  updates: AnimatedStyle<any>,
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  animatedValues: RefObject<AnimatedStyle<any>>,
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  adapters: ((updates: AnimatedStyle<any>) => void)[]
-): void => {
-  adapters.forEach((adapter) => {
-    adapter(updates);
-  });
-  animatedValues.current.value = {
-    ...animatedValues.current.value,
-    ...updates,
-  };
-
-  updateProps(viewDescriptors, updates);
-};
+export const updatePropsJestWrapper = makeUpdatePropsJestWrapper(updateProps);
 
 export default updateProps;
 
-// is-tree-shakable-suppress
-const maybeThrowError = () => {
-  // Jest attempts to access a property of this object to check if it is a Jest mock
-  // so we can't throw an error in the getter.
-  if (!IS_JEST) {
-    throw new Error(
-      '[Reanimated] `UpdatePropsManager` is not available on non-native platform.'
-    );
-  }
-};
 global.UpdatePropsManager = new Proxy(
   {},
   {
@@ -70,11 +39,3 @@ global.UpdatePropsManager = new Proxy(
     },
   }
 );
-
-/**
- * This used to be `SharedValue<Descriptors[]>` but objects holding just a
- * single `value` prop are fine too.
- */
-interface ViewDescriptorsWrapper {
-  value: Readonly<Descriptor[]>;
-}
