+++ packages/react-native-reanimated/src/updateProps/updatePropsBase.ts
@@ -1,0 +1,55 @@
+'use strict';
+
+import type { RefObject } from 'react';
+
+import { IS_JEST } from '../common';
+import type { AnimatedStyle } from '../commonTypes';
+import type { PropUpdates } from '../createAnimatedComponent/commonTypes';
+import type { Descriptor } from '../hook/commonTypes';
+
+/**
+ * This used to be `SharedValue<Descriptors[]>` but objects holding just a
+ * single `value` prop are fine too.
+ */
+export interface ViewDescriptorsWrapper {
+  value: Readonly<Descriptor[]>;
+}
+
+type UpdateProps = (
+  viewDescriptors: ViewDescriptorsWrapper,
+  updates: PropUpdates,
+  isAnimatedProps?: boolean
+) => void;
+
+export function makeUpdatePropsJestWrapper(updateProps: UpdateProps) {
+  return (
+    viewDescriptors: ViewDescriptorsWrapper,
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
+    updates: AnimatedStyle<any>,
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
+    animatedValues: RefObject<AnimatedStyle<any>>,
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
+    adapters: ((updates: AnimatedStyle<any>) => void)[]
+  ): void => {
+    adapters.forEach((adapter) => {
+      adapter(updates);
+    });
+    animatedValues.current.value = {
+      ...animatedValues.current.value,
+      ...updates,
+    };
+
+    updateProps(viewDescriptors, updates);
+  };
+}
+
+// is-tree-shakable-suppress
+export const maybeThrowError = () => {
+  // Jest attempts to access a property of this object to check if it is a Jest mock
+  // so we can't throw an error in the getter.
+  if (!IS_JEST) {
+    throw new Error(
+      '[Reanimated] `UpdatePropsManager` is not available on non-native platform.'
+    );
+  }
+};
